requirements.txt
ws4py==0.4.2
wheel==0.29.0
setuptools==36.0.1
python-aiml==0.9.0
pyflakes==1.5.0
pycodestyle==2.3.1
pip==9.0.1
mccabe==0.6.1
flake8==3.3.0
configparser==3.5.0
